Here's the file 'Trump-Harris.csv'
Well, i'll talk about the process of cleaning 'trump.csv' and 'harris.csv' file in here.

First of all, when you read this, you're suppose to take a look at the file 'trump.csv' or 'harris.csv' to understand the data type and what it tells.
Then, you'll notice that even we download the data from facebook library and it's all structured data but many columns are not in the right format to analyze in the analytic step.
So having an idea how we can transform these data for the analytic step is very important. This step is kinda difficult for me and i had struggle on this, especially on the column 'demographic_distribution' and 'delivery_by_region'. What about you? Do you feel the same? Do you have any ideas to work with these issues?
Okay, if you already read the file and have some imagine about what you are gonna do, let get into it.

First i'll combine 2 csv file because i want to make a compare the difference between Trump and Harris, both 2 files have the same data structure so it's very easy to combine them. 
After that i take a look at Null value and think what i should do with them. Well, in this case most of the time i leave the null value there and do nothing with them.

Then i'll remove some column that i feel unnecessary. It's 'languages', 'publisher_platforms' and 'currency'
    languages : Calling for election is calling for American residents on American soil. English is essential.
    publisher_platforms: The taget of this project is Advertise from Facebook platform and i download it from Facebook, the info about instagram doestn't matter.
    currency: all is USD, so i just care about then spend amount, not currency.